+++ be/src/pipeline/local_exchange/local_exchanger.cpp
@@ -25,7 +25,7 @@
 #include "vec/runtime/partitioner.h"
 
 namespace doris::pipeline {
-
+#include "common/compile_check_begin.h"
 template <typename BlockType>
 void Exchanger<BlockType>::_enqueue_data_and_set_ready(int channel_id,
                                                        LocalExchangeSinkLocalState& local_state,
@@ -170,11 +170,11 @@ Status ShuffleExchanger::get_block(RuntimeState* state, vectorized::Block* block
 Status ShuffleExchanger::_split_rows(RuntimeState* state, const uint32_t* __restrict channel_ids,
                                      vectorized::Block* block,
                                      LocalExchangeSinkLocalState& local_state) {
-    const auto rows = block->rows();
+    const auto rows = int32_t(block->rows());
     auto row_idx = std::make_shared<vectorized::PODArray<uint32_t>>(rows);
     {
         local_state._partition_rows_histogram.assign(_num_partitions + 1, 0);
-        for (size_t i = 0; i < rows; ++i) {
+        for (int32_t i = 0; i < rows; ++i) {
             local_state._partition_rows_histogram[channel_ids[i]]++;
         }
         for (int32_t i = 1; i <= _num_partitions; ++i) {
@@ -212,7 +212,7 @@ Status ShuffleExchanger::_split_rows(RuntimeState* state, const uint32_t* __rest
          */
         const auto& map = local_state._parent->cast<LocalExchangeSinkOperatorX>()
                                   ._shuffle_idx_to_instance_idx;
-        new_block_wrapper->ref(map.size());
+        new_block_wrapper->ref(int(map.size()));
         for (const auto& it : map) {
             DCHECK(it.second >= 0 && it.second < _num_partitions)
                     << it.first << " : " << it.second << " " << _num_partitions;
@@ -241,7 +241,7 @@ Status ShuffleExchanger::_split_rows(RuntimeState* state, const uint32_t* __rest
     } else {
         DCHECK(!bucket_seq_to_instance_idx.empty());
         new_block_wrapper->ref(_num_partitions);
-        for (size_t i = 0; i < _num_partitions; i++) {
+        for (int i = 0; i < _num_partitions; i++) {
             uint32_t start = local_state._partition_rows_histogram[i];
             uint32_t size = local_state._partition_rows_histogram[i + 1] - start;
             if (size > 0) {
@@ -426,7 +426,7 @@ Status BroadcastExchanger::sink(RuntimeState* state, vectorized::Block* in_block
     local_state._shared_state->add_total_mem_usage(wrapper->data_block.allocated_bytes(),
                                                    local_state._channel_id);
     wrapper->ref(_num_partitions);
-    for (size_t i = 0; i < _num_partitions; i++) {
+    for (int i = 0; i < _num_partitions; i++) {
         _enqueue_data_and_set_ready(i, local_state, {wrapper, {0, wrapper->data_block.rows()}});
     }
 
@@ -500,11 +500,11 @@ Status AdaptivePassthroughExchanger::_split_rows(RuntimeState* state,
                                                  const uint32_t* __restrict channel_ids,
                                                  vectorized::Block* block,
                                                  LocalExchangeSinkLocalState& local_state) {
-    const auto rows = block->rows();
+    const auto rows = int32_t(block->rows());
     auto row_idx = std::make_shared<std::vector<uint32_t>>(rows);
     {
         local_state._partition_rows_histogram.assign(_num_partitions + 1, 0);
-        for (size_t i = 0; i < rows; ++i) {
+        for (int32_t i = 0; i < rows; ++i) {
             local_state._partition_rows_histogram[channel_ids[i]]++;
         }
         for (int32_t i = 1; i <= _num_partitions; ++i) {
@@ -517,7 +517,7 @@ Status AdaptivePassthroughExchanger::_split_rows(RuntimeState* state,
             local_state._partition_rows_histogram[channel_ids[i]]--;
         }
     }
-    for (size_t i = 0; i < _num_partitions; i++) {
+    for (int32_t i = 0; i < _num_partitions; i++) {
         const size_t start = local_state._partition_rows_histogram[i];
         const size_t size = local_state._partition_rows_histogram[i + 1] - start;
         if (size > 0) {
